+++ src/ng-app/app/profile-details-edit/profile-details-edit.component.ts
@@ -135,6 +135,7 @@ export class ProfileDetailsEditComponent implements OnInit, OnDestroy {
     private tdpLabels: Map<string, string>;
 
     public showFanGraphs = false;
+    public showTGPChart = false;
 
     public infoTooltipShowDelay = 700;
 
@@ -143,9 +144,15 @@ export class ProfileDetailsEditComponent implements OnInit, OnDestroy {
 
     public fansOffAvailable = true;
 
+    public nvidiaPowerCTRLDefaultPowerLimit: number = 0;
+    public nvidiaPowerCTRLMaxPowerLimit: number = 1000;
+    public nvidiaPowerCTRLAvailable: boolean = false;
+
     public tempCustomFanCurve: ITccFanProfile = undefined;
 
     public get hasMaxFreqWorkaround() { return this.compat.hasMissingMaxFreqBoostWorkaround; }
+
+    public min = Math.min;
 
     @ViewChild('inputName') inputName: MatInput;
 
@@ -246,6 +253,24 @@ export class ProfileDetailsEditComponent implements OnInit, OnDestroy {
         this.subscriptions.add(this.tccDBus.isX11.subscribe(nextIsX11 => {
             if (nextIsX11 !== this.isX11) {
                 this.isX11 = nextIsX11;
+            }
+        }));
+
+        this.subscriptions.add(this.tccDBus.nvidiaPowerCTRLDefaultPowerLimit.subscribe(nextNVIDIAPowerCTRLDefaultPowerLimit => {
+            if (nextNVIDIAPowerCTRLDefaultPowerLimit !== undefined && nextNVIDIAPowerCTRLDefaultPowerLimit !== this.nvidiaPowerCTRLDefaultPowerLimit) {
+                this.nvidiaPowerCTRLDefaultPowerLimit = nextNVIDIAPowerCTRLDefaultPowerLimit;
+            }
+        }));
+
+        this.subscriptions.add(this.tccDBus.nvidiaPowerCTRLMaxPowerLimit.subscribe(nextNVIDIAPowerCTRLMaxPowerLimit => {
+            if (nextNVIDIAPowerCTRLMaxPowerLimit !== undefined && nextNVIDIAPowerCTRLMaxPowerLimit !== this.nvidiaPowerCTRLMaxPowerLimit) {
+                this.nvidiaPowerCTRLMaxPowerLimit = nextNVIDIAPowerCTRLMaxPowerLimit;
+            }
+        }));
+
+        this.subscriptions.add(this.tccDBus.nvidiaPowerCTRLAvailable.subscribe(nextNVIDIAPowerCTRLAvailable => {
+            if (nextNVIDIAPowerCTRLAvailable !== this.nvidiaPowerCTRLAvailable) {
+                this.nvidiaPowerCTRLAvailable = nextNVIDIAPowerCTRLAvailable;
             }
         }));
 
@@ -413,6 +438,8 @@ export class ProfileDetailsEditComponent implements OnInit, OnDestroy {
             tdpValues: odmTDPValuesArray
         });
 
+        const nvidiaPowerCTRLProfileGroup: FormGroup = this.fb.group(profile.nvidiaPowerCTRLProfile);
+
         cpuGroup.controls.scalingMinFrequency.setValidators([maxControlValidator(cpuGroup.controls.scalingMaxFrequency)]);
         cpuGroup.controls.scalingMaxFrequency.setValidators([minControlValidator(cpuGroup.controls.scalingMinFrequency)]);
 
@@ -432,7 +459,8 @@ export class ProfileDetailsEditComponent implements OnInit, OnDestroy {
             webcam: webcamGroup,
             fan: fanControlGroup,
             odmProfile: odmProfileGroup,
-            odmPowerLimits: odmPowerLimits
+            odmPowerLimits: odmPowerLimits,
+            nvidiaPowerCTRLProfile: nvidiaPowerCTRLProfileGroup
         });
 
         fg.controls.name.setValidators([Validators.required, Validators.minLength(1), Validators.maxLength(50)]);
@@ -804,6 +832,17 @@ export class ProfileDetailsEditComponent implements OnInit, OnDestroy {
             this.scrollTo.emit(this.fancontrolHeaderE.nativeElement.offsetTop - 50);
         } else {
             this.showFanGraphs = false;
+        }
+    }
+
+    @ViewChild('nvidiaPowerCTRLHeader') nvidiaPowerCTRLHeaderE;
+    public toggleTGPChart() {
+        this.showTGPChart = !this.showTGPChart;
+        if (this.showTGPChart) {
+            // The timeout is required here, because this is at the bottom of
+            // the page and the scroll needs to happen after the page already
+            // got rendered and is already bigger.
+            setTimeout(() => this.scrollTo.emit(this.nvidiaPowerCTRLHeaderE.nativeElement.offsetTop - 50), 10);
         }
     }